pyproject.toml
[tool.poetry]
name = "shift-left-secure"
version = "0.1.0"
description = "Mitigate vulnerabilities before pushing code to github/gitlab/bitbucket"
authors = ["Dhrumil Mistry <56185972+dmdhrumilmistry@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "shift_left_secure"}]

[tool.poetry.dependencies]
python = "^3.7"
openapi = "^1.1.0"
gitpython = "3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
